chore(weave_ts): Pull out openAIResponse test helper
